requirements.txt
langchain 
langchain-community 
chromadb
langchain-openai
streamlit
python-dotenv
langchain-chroma
pypdf
PyPDF2